Refactor: Improve component comments and organize imports
- Updated heading comments for App and AppRoutes components for better clarity and structure.

- Organized imports in App.tsx and index.tsx for better readability.

- Added small comments to explain specific functionality within components.
